combined user folder into db/user.js
